chore: refactor and minify diff
