+++ lib/vocabularies/code.ts
@@ -5,30 +5,7 @@ import {CodeGen, _, and, or, not, nil, strConcat, getProperty, Code, Name} from 
 import {alwaysValidSchema, Type} from "../compile/util"
 import N from "../compile/names"
 import {useFunc} from "../compile/util"
-
-export function getSkipCondition(cxt: KeywordCxt, prop: string): Code | undefined {
-  const propSchema = cxt.parentSchema.properties?.[prop]
-  if (!propSchema) return undefined
-
-  const hasReadOnly = propSchema.readOnly === true
-  const hasWriteOnly = propSchema.writeOnly === true
-
-  if (!hasReadOnly && !hasWriteOnly) return undefined
-
-  const conditions: Code[] = []
-  const oasContext = _`typeof ${N.this} == "object" && ${N.this} && ${N.this}.oas`
-
-  if (hasReadOnly) {
-    conditions.push(_`${oasContext} && ${N.this}.oas.mode === "request"`)
-  }
-
-  if (hasWriteOnly) {
-    conditions.push(_`${oasContext} && ${N.this}.oas.mode === "response"`)
-  }
-
-  return conditions.length === 1 ? conditions[0] : or(...conditions)
-}
-
+import {getSkipCondition} from "./oasContext"
 export function checkReportMissingProp(cxt: KeywordCxt, prop: string): void {
   const {gen, data, it} = cxt
   gen.if(noPropertyInData(gen, data, prop, it.opts.ownProperties), () => {
@@ -37,16 +14,15 @@ export function checkReportMissingProp(cxt: KeywordCxt, prop: string): void {
   })
 }
 
-export function checkMissingProp(cxt: KeywordCxt, properties: string[], missing: Name): Code {
-  const {
-    gen,
-    data,
-    it: {opts},
-  } = cxt
+export function checkMissingProp(
+  {gen, data, it: {opts}, parentSchema}: KeywordCxt,
+  properties: string[],
+  missing: Name
+): Code {
   return or(
     ...properties.map((prop) =>
       and(
-        not(getSkipCondition(cxt, prop) ?? _`false`),
+        not(getSkipCondition(parentSchema, prop) ?? _`false`),
         noPropertyInData(gen, data, prop, opts.ownProperties),
         _`${missing} = ${prop}`
       )
