+++ lib/vocabularies/validation/required.ts
@@ -5,9 +5,9 @@ import {
   reportMissingProp,
   propertyInData,
   noPropertyInData,
-  getSkipCondition,
   checkMissingProp,
 } from "../code"
+import {getSkipCondition} from "../oasContext"
 import {_, str, nil, not, Name, Code} from "../../compile/codegen"
 import {checkStrictMode} from "../../compile/util"
 
@@ -53,7 +53,7 @@ const def: CodeKeywordDefinition = {
         cxt.block$data(nil, loopAllRequired)
       } else {
         for (const prop of schema) {
-          const skip = getSkipCondition(cxt, prop) ?? _`false`
+          const skip = getSkipCondition(cxt.parentSchema, prop) ?? _`false`
           /**
            * Generate a runtime check: validate `required` only when this property
            * should NOT be skipped in the current context (readOnly/writeOnly).
